requirements.txt
mergedeep
numpy
opencv-python
pyserial
pytesseract
pyyaml
telegram-send
